Support TypeScript upgrade tests (#19)
### TypeScript programs are now supported for upgrade tests

With this change, upgrade tests can work with programs written in TypeScript and not just YAML. This is important for quickly adding coverage to existing providers like pulumi-eks and pulumi-awsx without writing new programs. The internals currently lean heavily into ProgramTest machinery to make sure that the baseline<->SUT transition correctly reinstalls the TypeScript SDK for the provider as well as the provider itself at the expected versions.

### Upgrade tests support additional baseline dependencies

WithExtraBaselineDependencies is added to support the pulumi-eks use case. This provider is tightly integrated with the AWS provider, and needs to be able to record baselines on a pinned version of itself as well as a pinned version of the AWS provider. This is now made possible.

### Upgrade test for YAML programs are now supported without pinning the provider

Previously when testing YAML programs for upgrades, the user was expected to specify the explicit provider at the baseline version in the YAML definitions. This was cumbersome and required cross-checking the baseline version given to the test framework against the baseline version written in the YAML program. This restriction is now lifted in favor of the test framework installing the expected plugin versions as ambient plugins in PATH. WithExtraBaselineDependencies are also supported now.


---------

Co-authored-by: Ian Wahbe <ian@wahbe.com>