Implements more precise diff validation (#30)
Fixes #24 

For PreviewOnly upgrade tests, the user is now able to write custom validations over what changes are permitted or not. This relies on inspecting Diff method logs. Two canned strategies are provided, NoChanges which is the default, and NoReplacements which is more lenient. Custom code can narrow this down further, for example only prohibiting replacements by type.