Retain temporary folders for inspection when tests failed (#81)
* Retain temporary folders for inspection when tests failed

This came out of the friction of debugging a failed upgrade test. ProgramTest had a nice feature that left a folder in
TMP with the current state right at the point of failure that can be further inspected in the CLI, printing the path to
the folder when that happens.

This change moves us closer to the desired experience by substituting t.TempDir() with an inlined version of TempDir
that does not delete it on failure; and similarly it stops destroying stacks for failed tests.

There is some further ergonomics to work out such as emitting an .envrc file so that all the env vars get picked up in
the dir and the user does not need to chase down the file backend URL, passphrase and so on.

Refining this on PR feedback, cleanup is NOT suspended in CI environments. Only in local environments. This makes sure in particular that Pulumi still calls `destroy` to clean up resources in test accounts that can be expensive.