Clean up failed tests by default
Fixes #109

A common developer experience issue is that which iterating locally, when their test fails, they have to go and manually delete created resources. This changes the behaviour to make it an explicit opt-in to leave the resources. It will, however continue to default to keeping the temp directories on failure.