Bump minimum pulumi requirement (#119)
https://github.com/pulumi/pulumi/pull/17898 was a breaking change for
projects that directly depend on `pulumi/sdk` but not `pulumi/pkg`,
which is the case with several native providers.

If these projects upgrade to `sdk@v3.143` or later, they can still
indirectly depend on earlier versions of `pkg` via libraries like this
one. This leads to errors like the following:

```
go: github.com/pulumi/pulumi-command/provider/pkg/provider tested by
	github.com/pulumi/pulumi-command/provider/pkg/provider.test imports
	github.com/pulumi/providertest imports
	github.com/pulumi/pulumi/pkg/v3/testing/integration imports
	github.com/pulumi/pulumi/sdk/v3/go/common/tools: cannot find module providing package github.com/pulumi/pulumi/sdk/v3/go/common/tools
```

This bumps our minimum required version of `pkg` to one which accounts
for https://github.com/pulumi/pulumi/pull/17898.

This problem is self-inflicted and only exists because we choose to
version `sdk` and `pkg` separately despite there being a tight coupling
between the two. We would not have this problem if we released them
together as a single module with `sdk` and `pkg` packages.