Show diffs for all previews (#120)
Currently, a failing preview in a test just shows as something like
```
azure-native:storage:Blob myBlobFile create replacement [diff: ~source]
```

which is not as helpful as it could be diagnose the problem because you
don't know _what_ the diff is.

We could expose a flag for an optional diff in `optproviderupgrade` but
since this is a testing library, I suggest we skip this boilerplate and
just turn the diff always on - when would you _not_ want to see the diff
on test failure?

With this change:
```
            +-azure-native:storage:Blob: (replace)
                [id=/subscriptions/0282681f-7a9e-424b-80b2-96babd57a8a1/resourceGroups/rg1c1afbc6/providers/Microsoft.Storage/storageAccounts/store16dfec76/blobServices/default/containers/assets/blobs/mon-fichier.txt]
                [urn=urn:pulumi:test::upgrade-storage-blob::azure-native:storage:Blob::myBlobFile]
              - source: [secret]
              + source: asset(text:678e7ad) {
                    <contents elided>
                }
```