Update README.md to steer toward pulumitest
This repo is weird in that there's essentially a new and improved version in the pulumitest directory, but the older providertest code is still available in the top level (and used in some older providers tests). 

We need to clarify this in the README, so that users (internal and external) don't end up trying to write more tests that depend on providertest rather than pulumitest.