Fix retention & location of temp dir via env vars (#126)
## Part 1: Allow using ENV vars for temp directory control

- Add specific environment variable to always retain directories and fix
instruction that didn't work for passing tests to use the new variable.
- Document environment variables in readme and document decision process
for retaining directories in code comments.
- Allow setting a default temp dir via env var - this would allow it to
be configured more easily in CI so that we can grab failed test
directories as workflow assets.

Example config file to use this locally when running in VSCode:

```json
{
  "go.testTimeout": "320s",
  "go.testEnvVars": {
    "PULUMITEST_TEMP_DIR": "${workspaceFolder}/.tmp",
  }
}
```

## Part 2: Run with custom temp dir in CI & capture failures

This highlighted that we need to make the initial directory creation
safe (added a mutex & rewrote a bad test). This also sets up an example
of how to execute tests in CI and capture the temp directory on failure.

GitHub actions example:

```yaml
      - name: Run acceptance tests
        run: go test
        env:
          PULUMITEST_TEMP_DIR: ${{ github.workspace }}/test_temp
          PULUMITEST_RETAIN_FILES_ON_FAILURE: "true"

      - name: Upload test_temp directory on failure
        if: failure()
        uses: actions/upload-artifact@v4
        with:
          path: test_temp
```

Example test run with failing test:
https://github.com/pulumi/providertest/actions/runs/13902023819?pr=127